% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmer.R
\name{greta.glmer}
\alias{greta.glmer}
\title{Initialize the linear predictor for generalized linear effect models}
\usage{
greta.glmer(formula, data, prior_intercept = NULL,
  prior_coefficients = NULL, prior_random_effects = NULL)
}
\arguments{
\item{formula}{an \code{lme4}-style formula}

\item{data}{the data set containing the variables described in
\code{formula}}

\item{prior_intercept}{a \code{\link{greta::greta_array}} random variable
that specifies the prior for the intercept \eqn{\beta_0} for the
fixed effects design
matrix. If \code{NULL}, the distribution for the intercept is chosen as
\deqn{p(\beta_0) \propto const}. The intercept prior
needs to have dimension 1.}

\item{prior_coefficients}{a \code{\link{greta::greta_array}} random
variable that specifies the prior for the coefficients\eqn{\beta} for the
fixed effects
design matrix. If \code{NULL}, the distribution for the coefficients is
chosen as \deqn{p(\beta_i) \sim \mathcal{N}(0, \sigma_{\beta})}, and
 \deqn{p(\sigma_{\beta}) \sim \text{Half-Cauchy}(0, \inf)}.
The coefficients prior
need to have the same dimensionalty as the number of columns of
your fixed effects design matrix (without the intercept).}

\item{prior_random_effects}{a \code{\link{greta::greta_array}} random
variable that specifies the prior for the random effects \eqn{\beta} for
the random effects design matrix. If \code{NULL}, the distribution for the
 random effects is chosen as
 \deqn{p(\gamma_g) \sim \mathcal{N}(0, \sigma_{\gamma_g})}, and
 \deqn{p(\sigma_{\gamma_g}) \sim \text{Wishart}(\text{len}(\gamma_g) + 1, \text{diag}(\text{len}(\gamma_g)))}.
If \code{prior_random_effects} is provided
it needs to have the same dimensionalty as the
number of columns of \eqn{Z}
random effect design matrix. These are generally not trivial to construct.
Calling \code{lme4::glFormula(formula, data)} creates the specified
random effects matrix \eqn{Z}. From this the dimensionality of
\code{prior_random_effects} can be inferred.}
}
\value{
Returns a list with the following elements:
\item{}{}
}
\description{
\code{greta.glmer} creates the linear predictor $eta$ required
for fitting generalized linear mixed models. The linear predictor has the
form: \deqn{\eta = X \beta + Z \gamma},
where \eqn{X} is the fixed effects design matrix, \eqn{\beta} the fixed
effects, \eqn{Z} the random effects design matrix and \eqn{\gamma} the
random effects.
}
\examples{

greta.glmer(Sepal.Length ~ Sepal.Width + (1 | Species), iris)
}
